usm: http2: Move a test utility to a test-only-file
